pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "socialos"
version = "2.0.0"
description = "SocialOS — Production-Grade Social Media Management Platform"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.12"
dependencies = [
    # Core Django
    "django>=5.0,<6.0",
    # REST Framework
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "drf-spectacular>=0.27,<1.0",
    "django-filter>=24.0,<25.0",
    "django-cors-headers>=4.3,<5.0",
    # Real-time / WebSockets
    "channels>=4.1,<5.0",
    "channels-redis>=4.2,<5.0",
    "daphne>=4.1,<5.0",
    # Background Tasks
    "celery[redis]>=5.3,<6.0",
    "django-celery-beat>=2.6,<3.0",
    "django-celery-results>=2.5,<3.0",
    "celery-redbeat>=2.2,<3.0",
    "redis>=5.0,<6.0",
    # Database
    "psycopg2-binary>=2.9,<3.0",
    "django-pgcrypto>=1.4,<2.0",
    # Storage & Media
    "django-storages[s3]>=1.14,<2.0",
    "Pillow>=10.3,<11.0",
    # Auth & Security
    "social-auth-app-django>=5.4,<6.0",
    "django-axes>=6.4,<7.0",
    "cryptography>=42.0,<43.0",
    # Performance & Caching
    "django-cachalot>=2.6,<3.0",
    "whitenoise>=6.6,<7.0",
    # HTTP Client (async, for platform API calls)
    "httpx>=0.27,<1.0",
    # Configuration
    "python-decouple>=3.8,<4.0",
    # Monitoring & Error Tracking
    "sentry-sdk[django]>=2.0",
    # ASGI Server
    "gunicorn>=22.0,<23.0",
    "uvicorn[standard]>=0.30,<1.0",
    # Event Streaming (Kafka)
    "confluent-kafka>=2.4,<3.0",
    # AI Clients
    "openai>=1.30,<2.0",
    "anthropic>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
    # Profiling & Debug
    "django-silk>=5.2,<6.0",
    "django-extensions>=3.2,<4.0",
    "Werkzeug>=3.0",          # Required by runserver_plus

    # Testing
    "factory-boy>=3.3,<4.0",
    "pytest>=8.2,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "faker>=25.0",

    # Type Checking
    "mypy>=1.10,<2.0",
    "django-stubs[compatible-mypy]>=5.0",
    "djangorestframework-stubs>=3.15",

    # Security Scanning
    "bandit>=1.7,<2.0",
    "safety>=3.0",

    # Interactive shell
    "ipython>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["socialos", "apps", "utils"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "socialos.settings.development"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test", "*Tests"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring live services",
    "unit: marks tests as pure unit tests (no DB, no network)",
]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "socialos.settings.development"

# ---------------------------------------------------------------------------
# bandit (security linter)
# ---------------------------------------------------------------------------
[tool.bandit]
exclude_dirs = ["tests", "migrations", ".venv"]
skips = ["B101"]   # assert statements are fine in tests

# ---------------------------------------------------------------------------
# ruff (fast linter / formatter — install separately if needed)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]
exclude = ["migrations", ".venv"]

# ---------------------------------------------------------------------------
# isort (import sorting — compatible with black)
# ---------------------------------------------------------------------------
[tool.isort]
profile = "black"
line_length = 120
skip = [".venv"]
known_django = ["django"]
known_drf = ["rest_framework"]
known_first_party = ["apps", "socialos", "utils"]
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "DRF",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

# ---------------------------------------------------------------------------
# black (code formatter)
# ---------------------------------------------------------------------------
[tool.black]
line-length = 120
target-version = ["py312"]
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
)/
'''

# ---------------------------------------------------------------------------
# flake8 tool
# ---------------------------------------------------------------------------
[tool.flake8]
max-line-length = 120
exclude = [
    "__pycache__",
    ".venv",
]
